backend/requirements.txt
Django==5.1.6
psycopg2==2.9.10
django-environ==0.12.0
django-cors-headers==4.7.0

djoser==2.3.1

djangorestframework==3.15.2
Markdown==3.7
django-filter==25.1
drf-spectacular==0.28.0

django-simple-history==3.8.0